fix: move results panel into right inspector (#240)
